cargo-release-plan: delegate the requirement predicate to its shared definition
Extracting the predicate into manifest.rs left the original copy behind in
check.rs, so the module documented as the single definition had a second one
beside it. The copy had already drifted: it neither trimmed its input nor
accepted the bare spelling, which is the form apply reads from manifest text.

Copilot-Session: c188591e-e511-43b5-a6ec-e3ed0d74d39b
